requirements.txt
streamlit
streamlit-autorefresh
nba_api
requests
python-dotenv
